requirements.txt
# Core backend
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sqlalchemy==2.0.31
pymysql==1.1.1
pandas==2.2.3
tqdm==4.66.5

# LangChain + Gemini stack
langchain==0.3.2
langchain-core==0.3.9
langchain-community==0.3.1
langchain-google-genai==1.0.3
google-generativeai==0.8.2

# NLP & vector search
faiss-cpu==1.8.0
sentence-transformers==2.2.2

# PDF and text extraction
pdfplumber==0.11.0
regex==2024.9.11

# Data validation
pydantic==1.10.15
